+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Preferences not saving on macOS
+- Keys staying held after window loses focus
 - Holding a key no longer twists the puzzle repeatedly
 
 ## [1.0.1] - 2022-11-23